Fixed links

+++ compile.py
@@ -10,7 +10,8 @@ output_directory = Path("docs")
 template = Path("_pages/_template.html")
 
 for old_output in output_directory.glob("*"):
-    old_output.unlink()
+    if old_output.is_file():
+        old_output.unlink()
 
 with template.open() as f:
     template = f.read()
